# frontend/swagger/be.yml
openapi: 3.1.0
info:
  title: edu-max BE
  version: 0.0.1
paths:
  /chats:
    get:
      tags:
        - chats
      summary: Get Chats
      description: 모든 채팅 목록을 반환합니다.
      operationId: get_chats_chats_get
      parameters:
        - name: subject
          in: query
          required: false
          schema:
            anyOf:
              - type: array
                items:
                  $ref: '#/components/schemas/SubjectEnum'
              - type: 'null'
            title: Subject
        - name: cursor
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Cursor
        - name: size
          in: query
          required: false
          schema:
            type: integer
            maximum: 100
            minimum: 1
            default: 10
            title: Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CursorPaginateResponse_ChatResponse_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
        - chats
      summary: Create Chat
      description: 새로운 채팅을 생성합니다.
      operationId: create_chat_chats_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatCreateRequestBody'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /problems:
    post:
      tags:
        - problems
      summary: Create Problems
      description: 문제 생성 API 입니다.
      operationId: create_problems_problems_post
      parameters:
        - name: x-api-key
          in: header
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: X-Api-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProblemCreateRequestBody'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemCreateResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /uploads:
    post:
      tags:
        - upload
      summary: Create Upload Url
      description: 파일 업로드 위한 url을 반환합니다.
      operationId: create_upload_url_uploads_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          description: 지원하지 않는 mime type 요청 시 반환됩니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /users:
    get:
      tags:
        - users
      summary: Get Users
      description: 전체 사용자 목록 조회
      operationId: get_users_users_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/UserResponse'
                type: array
                title: Response Get Users Users Get
      x-staleTime: Infinity
      x-gcTime: Infinity 
    post:
      tags:
        - users
      summary: Create User
      description: 신규 사용자 생성
      operationId: create_user_users_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequestBody'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserCreateResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /goals:
    post:
      tags:
        - goals
      summary: Create Goal
      description: 채팅 방에 맞는 목표를 생성합니다.
      operationId: create_goal_goals_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoalCreateRequestBody'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoalResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /problem-bookmarks:
    get:
      tags:
        - problem-bookmarks
      summary: Get Problem Bookmarks
      description: 사용자가 등록한 문제 학습 노트 리스트 조회
      operationId: get_problem_bookmarks_problem_bookmarks_get
      parameters:
        - name: subject
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/SubjectEnum'
        - name: status
          in: query
          required: false
          schema:
            type: array
            items:
              $ref: '#/components/schemas/ProblemStatus'
            default: []
            title: Status
        - name: cursor
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Cursor
        - name: size
          in: query
          required: false
          schema:
            type: integer
            maximum: 100
            minimum: 1
            default: 10
            title: Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: >-
                  #/components/schemas/CursorPaginateResponse_ProblemBookmarkResponse_
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
        - problem-bookmarks
      summary: Bookmark Problem
      description: 현재 채팅방의 문제를 북마크로 등록합니다.
      operationId: bookmark_problem_problem_bookmarks_patch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProblemBookmarkCreateRequestBody'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemBookmarkCreateResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ChatCreateRequestBody:
      properties:
        title:
          type: string
          title: Title
        grade:
          type: string
          title: Grade
        subject:
          type: string
          title: Subject
      type: object
      required:
        - title
        - grade
        - subject
      title: ChatCreateRequestBody
...
    ChatResponse:
      properties:
        id:
          type: integer
          title: Id
        title:
          type: string
          title: Title
        subject:
          type: string
          title: Subject
        has_problem:
          type: boolean
          title: Has Problem
        has_active_goal:
          type: boolean
          title: Has Active Goal
        latest_used_at:
          type: string
          format: date-time
          title: Latest Used At
      type: object
      required:
        - id
        - title
        - subject
        - has_problem
        - has_active_goal
        - latest_used_at
      title: ChatResponse
...
    ChoiceProblemContent:
      properties:
        subject:
          type: string
          title: Subject
        grade:
          type: integer
          title: Grade
        semester:
          anyOf:
            - type: integer
            - type: 'null'
          title: Semester
        level:
          type: integer
          title: Level
        primary:
          anyOf:
            - type: string
            - type: 'null'
          title: Primary
        secondary:
          anyOf:
            - type: string
            - type: 'null'
          title: Secondary
        specific:
          anyOf:
            - type: string
            - type: 'null'
          title: Specific
        category:
          anyOf:
            - type: string
            - type: 'null'
          title: Category
        problem:
          type: string
          title: Problem
        tags:
          anyOf:
            - items:
                type: string
              type: array
            - type: 'null'
          title: Tags
        hint:
          anyOf:
            - type: string
            - type: 'null'
          title: Hint
        explanation:
          anyOf:
            - type: string
            - type: 'null'
          title: Explanation
        choices:
          items:
            $ref: '#/components/schemas/ProblemChoice'
          type: array
          title: Choices
          default: []
      type: object
      required:
        - subject
        - grade
        - semester
        - level
        - primary
        - secondary
        - specific
        - category
        - problem
        - hint
        - explanation
      title: ChoiceProblemContent
    CursorPaginateResponse_ChatResponse_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ChatResponse'
          type: array
          title: Items
        next_cursor:
          anyOf:
            - type: string
            - type: 'null'
          title: Next Cursor
      type: object
      required:
        - items
        - next_cursor
      title: CursorPaginateResponse[ChatResponse]
    CursorPaginateResponse_ProblemBookmarkResponse_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ProblemBookmarkResponse'
          type: array
          title: Items
        next_cursor:
          anyOf:
            - type: string
            - type: 'null'
          title: Next Cursor
      type: object
      required:
        - items
        - next_cursor
      title: CursorPaginateResponse[ProblemBookmarkResponse]
...
    ErrorResponse:
      properties:
        message:
          type: string
          title: Message
        detail:
          anyOf:
            - additionalProperties: true
              type: object
            - type: 'null'
          title: Detail
      type: object
      required:
        - message
      title: ErrorResponse
    GoalCreateRequestBody:
      properties:
        chat_id:
          type: integer
          title: Chat Id
        goal_count:
          type: integer
          title: Goal Count
      type: object
      required:
        - chat_id
        - goal_count
      title: GoalCreateRequestBody
    GoalResponse:
      properties:
        id:
          type: integer
          title: Id
        goal_count:
          type: integer
          title: Goal Count
        solved_count:
          type: integer
          title: Solved Count
      type: object
      required:
        - id
        - goal_count
        - solved_count
      title: GoalResponse
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    MultipleChoiceProblemResponse:
      properties:
        id:
          type: string
          title: Id
        type:
          type: string
          const: 다중선택 객관식
          title: Type
        content:
          $ref: '#/components/schemas/ChoiceProblemContent'
      type: object
      required:
        - id
        - type
        - content
      title: MultipleChoiceProblemResponse
...
    MultipleShortAnswerProblemResponse:
      properties:
        id:
          type: string
          title: Id
        type:
          type: string
          const: 다중응답 주관식
          title: Type
        content:
          $ref: '#/components/schemas/ShortAnswerProblemContent'
      type: object
      required:
        - id
        - type
        - content
      title: MultipleShortAnswerProblemResponse
    ProblemBookmarkChatProblemResponse:
      properties:
        number:
          type: integer
          title: Number
        status:
          $ref: '#/components/schemas/ProblemStatus'
        last_answer:
          anyOf:
            - type: string
            - type: 'null'
          title: Last Answer
        problem_info:
          $ref: '#/components/schemas/ProblemResponse'
      type: object
      required:
        - number
        - status
        - last_answer
        - problem_info
      title: ProblemBookmarkChatProblemResponse
    ProblemBookmarkChatResponse:
      properties:
        id:
          type: integer
          title: Id
        has_active_goal:
          type: boolean
          title: Has Active Goal
      type: object
      required:
        - id
        - has_active_goal
      title: ProblemBookmarkChatResponse
...
    ProblemBookmarkCreateRequestBody:
      properties:
        chat_id:
          type: integer
          title: Chat Id
        problem_id:
          type: string
          title: Problem Id
        is_bookmarked:
          type: boolean
          title: Is Bookmarked
        bookmarked_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Bookmarked At
      type: object
      required:
        - chat_id
        - problem_id
        - is_bookmarked
      title: ProblemBookmarkCreateRequestBody
    ProblemBookmarkCreateResponse:
      properties:
        id:
          type: integer
          title: Id
        chat_id:
          type: integer
          title: Chat Id
        problem_id:
          type: string
          title: Problem Id
        is_bookmarked:
          type: boolean
          title: Is Bookmarked
        bookmarked_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Bookmarked At
      type: object
      required:
        - id
        - chat_id
        - problem_id
        - is_bookmarked
        - bookmarked_at
      title: ProblemBookmarkCreateResponse
    ProblemBookmarkResponse:
      properties:
        bookmarked_at:
          type: string
          format: date-time
          title: Bookmarked At
        chat:
          $ref: '#/components/schemas/ProblemBookmarkChatResponse'
        problem:
          $ref: '#/components/schemas/ProblemBookmarkChatProblemResponse'
      type: object
      required:
        - bookmarked_at
        - chat
        - problem
      title: ProblemBookmarkResponse
    ProblemChoice:
      properties:
        'no':
          type: integer
          title: 'No'
        inst:
          type: string
          title: Inst
      type: object
      required:
        - 'no'
        - inst
      title: ProblemChoice
    ProblemCreateAnswerRequest:
      properties:
        answer:
          type: string
          title: Answer
        accepted_answers:
          anyOf:
            - items:
                type: string
              type: array
            - type: 'null'
          title: Accepted Answers
      type: object
      required:
        - answer
      title: ProblemCreateAnswerRequest
    ProblemCreateChoiceRequest:
      properties:
        'no':
          type: integer
          title: 'No'
        inst:
          type: string
          title: Inst
      type: object
      required:
        - 'no'
        - inst
      title: ProblemCreateChoiceRequest
    ProblemCreateRequest:
      properties:
        id:
          type: string
          title: Id
        subject:
          type: string
          title: Subject
        grade:
          type: integer
          title: Grade
        type:
          type: string
          title: Type
        level:
          type: integer
          title: Level
        url:
          anyOf:
            - type: string
            - type: 'null'
          title: Url
        semester:
          anyOf:
            - type: integer
            - type: 'null'
          title: Semester
        primary:
          anyOf:
            - type: string
            - type: 'null'
          title: Primary
        secondary:
          anyOf:
            - type: string
            - type: 'null'
          title: Secondary
        specific:
          anyOf:
            - type: string
            - type: 'null'
          title: Specific
        category:
          anyOf:
            - type: string
            - type: 'null'
          title: Category
        problem:
          type: string
          title: Problem
        choices:
          anyOf:
            - items:
                $ref: '#/components/schemas/ProblemCreateChoiceRequest'
              type: array
            - type: 'null'
          title: Choices
        correct_answers:
          items:
            $ref: '#/components/schemas/ProblemCreateAnswerRequest'
          type: array
          title: Correct Answers
        explanation:
          anyOf:
            - type: string
            - type: 'null'
          title: Explanation
        hint:
          anyOf:
            - type: string
            - type: 'null'
          title: Hint
        tags:
          anyOf:
            - items:
                type: string
              type: array
            - type: 'null'
          title: Tags
      type: object
      required:
        - id
        - subject
        - grade
        - type
        - level
        - url
        - semester
        - primary
        - secondary
        - specific
        - category
        - problem
        - choices
        - correct_answers
        - explanation
        - hint
        - tags
      title: ProblemCreateRequest
    ProblemCreateRequestBody:
      properties:
        problems:
          items:
            $ref: '#/components/schemas/ProblemCreateRequest'
          type: array
          title: Problems
      type: object
      required:
        - problems
      title: ProblemCreateRequestBody
    ProblemCreateResponse:
      properties:
        ids:
          items:
            type: string
          type: array
          title: Ids
      type: object
      required:
        - ids
      title: ProblemCreateResponse
...
    ProblemResponse:
      oneOf:
        - $ref: '#/components/schemas/SingleChoiceProblemResponse'
        - $ref: '#/components/schemas/MultipleChoiceProblemResponse'
        - $ref: '#/components/schemas/SingleShortAnswerProblemResponse'
        - $ref: '#/components/schemas/MultipleShortAnswerProblemResponse'
      title: ProblemResponse
      discriminator:
        propertyName: type
        mapping:
          다중선택 객관식: '#/components/schemas/MultipleChoiceProblemResponse'
          다중응답 주관식: '#/components/schemas/MultipleShortAnswerProblemResponse'
          단일선택 객관식: '#/components/schemas/SingleChoiceProblemResponse'
          단일응답 주관식: '#/components/schemas/SingleShortAnswerProblemResponse'
    ProblemStatus:
      type: string
      enum:
        - 풀지 않음
        - 정답
        - 오답
        - 복습 완료
      title: ProblemStatus
...
    ShortAnswerProblemContent:
      properties:
        subject:
          type: string
          title: Subject
        grade:
          type: integer
          title: Grade
        semester:
          anyOf:
            - type: integer
            - type: 'null'
          title: Semester
        level:
          type: integer
          title: Level
        primary:
          anyOf:
            - type: string
            - type: 'null'
          title: Primary
        secondary:
          anyOf:
            - type: string
            - type: 'null'
          title: Secondary
        specific:
          anyOf:
            - type: string
            - type: 'null'
          title: Specific
        category:
          anyOf:
            - type: string
            - type: 'null'
          title: Category
        problem:
          type: string
          title: Problem
        tags:
          anyOf:
            - items:
                type: string
              type: array
            - type: 'null'
          title: Tags
        hint:
          anyOf:
            - type: string
            - type: 'null'
          title: Hint
        explanation:
          anyOf:
            - type: string
            - type: 'null'
          title: Explanation
      type: object
      required:
        - subject
        - grade
        - semester
        - level
        - primary
        - secondary
        - specific
        - category
        - problem
        - hint
        - explanation
      title: ShortAnswerProblemContent
...
    SingleChoiceProblemResponse:
      properties:
        id:
          type: string
          title: Id
        type:
          type: string
          const: 단일선택 객관식
          title: Type
        content:
          $ref: '#/components/schemas/ChoiceProblemContent'
      type: object
      required:
        - id
        - type
        - content
      title: SingleChoiceProblemResponse
...
    SingleShortAnswerProblemResponse:
      properties:
        id:
          type: string
          title: Id
        type:
          type: string
          const: 단일응답 주관식
          title: Type
        content:
          $ref: '#/components/schemas/ShortAnswerProblemContent'
      type: object
      required:
        - id
        - type
        - content
      title: SingleShortAnswerProblemResponse
...
    SubjectEnum:
      type: string
      enum:
        - math
        - english
      title: SubjectEnum
...
    UploadUrlRequest:
      properties:
        mime_type:
          type: string
          title: Mime Type
      type: object
      required:
        - mime_type
      title: UploadUrlRequest
    UploadUrlResponse:
      properties:
        presigned_url:
          type: string
          title: Presigned Url
        file_url:
          type: string
          title: File Url
      type: object
      required:
        - presigned_url
        - file_url
      title: UploadUrlResponse
    UserCreateRequestBody:
      properties:
        name:
          type: string
          title: Name
        grade:
          type: integer
          title: Grade
        memo:
          anyOf:
            - type: string
            - type: 'null'
          title: Memo
      type: object
      required:
        - name
        - grade
      title: UserCreateRequestBody
    UserCreateResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        grade:
          type: integer
          title: Grade
        memo:
          anyOf:
            - type: string
            - type: 'null'
          title: Memo
      type: object
      required:
        - id
        - name
        - grade
        - memo
      title: UserCreateResponse
    UserResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        grade:
          type: integer
          title: Grade
        memo:
          anyOf:
            - type: string
            - type: 'null'
          title: Memo
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
        - id
        - name
        - grade
        - memo
        - created_at
        - updated_at
      title: UserResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError